requirements.txt
langchain
openai
chromadb
streamlit
python-dotenv
sqlvalidator
